controllers send 401 AND progress update

+++ controllers/EducationsController.js
@@ -20,6 +20,7 @@ export default class EducationsController {
 
       // Their particular username to store with the data
       let username = await validationService.validateUser(req);
+      if (username === "") return (res.status(401))
 
       // TODO:  if validateUser returns empty username, what to do
 
@@ -41,6 +42,7 @@ export default class EducationsController {
     try {
       // Their particular username to store with the data
       let username = await validationService.validateUser(req);
+      if (username === "") return (res.status(401))
 
       req.body.username = username;
       console.log(req.body)
@@ -61,6 +63,7 @@ export default class EducationsController {
     try {
       // Their particular username to store with the data
       let thisUser = await validationService.validateUser(req);
+      if (thisUser === "") return (res.status(401))
 
       await dbContext.Education.findByIdAndRemove({ username: thisUser, _id: req.params.educationId });
